requirements.txt
Django==4.0.6
celery==5.2.7 
psycopg2-binary==2.9.3
redis==4.3.4
python3-linkedin
djangorestframework==3.13.1
